:recycle: Change Cache interface position
